Interface of manager
